% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_energy_demand.R
\name{calculate_energy_demand}
\alias{calculate_energy_demand}
\title{Calculate daily energy requirements}
\usage{
calculate_energy_demand(n, bodymass, metabolism = "FMR",
  assimilation = 0.73, adjust = NULL, t1, t2, plot = TRUE)
}
\arguments{
\item{n}{Number of individuals in the population; vector of values for each time step}

\item{bodymass}{Average body mass (kg) of individuals in the population; vector of values for each time step}

\item{metabolism}{Type of metabolic rate to use: \code{"FMR"} for field metabolic rate or \code{"RMR"} for resting metabolic rate; see Details}

\item{assimilation}{Estimated assimilation efficiency of energy consumed; defaults to 0.73 estimated for shorebirds (Castro et al. 1989)}

\item{adjust}{Optional; allows manual adjustment of DER between time steps \code{t1} and \code{t2} by adding a specified percentage of the calculated DER; see Details}

\item{t1}{Optional; starting time step for adjustment}

\item{t2}{Optional; ending time step for adjustment}

\item{plot}{Logical; if \code{TRUE}, also returns a plot of daily energy requirements for convenience}
}
\value{
Returns vector of daily energy requirements (kJ) for the population at each time step.
}
\description{
Calculates total daily energy requirement (DER) for a population based on an allometric scaling equation developed for shorebirds (Kersten and Piersma 1987).
}
\details{
Daily energy requirements (DER) for a population are calculated based on allometric scaling equations, requiring only an estimate of average body mass (kg) and the number of individuals in the population. The function includes four equations:

(1) RMR = 437*m^0.729, where RMR is resting metabolic rate and m is body mass (kg), derived for shorebirds (Kersten and Piersma 1987)

(2) FMR = 3 * RMR, where FMR is field metabolic rate, based on comparisons of lab and field animals, and taking into account additional energy expended due to daily activities and thermoregulation (Kersten and Piersma 1987)

(3) DEI = FMR/assimilation efficiency, where DEI is daily energy intake, because not all energy consumed is assimilated; if RMR is selected, equation is instead DEI = RMR/assimilation efficiency

(4) DER = DEI * n, where n is the number of individuals in the population

To manually adjust the calculated DER up or down during certain time steps, e.g. increasing by a percentage prior to a migration window, use the adjust parameter to specify the percentage change (values between 0 and 1) and specify the time steps.
}
\examples{
calculate_energy_demand(n=c(100,150,200), bodymass=c(0.5,0.6,0.7),
   metabolism='FMR', assimilation=0.73)
calculate_energy_demand(n=c(100,150,200), bodymass=c(0.5,0.6,0.7),
   metabolism='FMR', assimilation=0.73, adjust=0.25, t1=3, t2=3)

}
\author{
Kristen Dybala, \email{kdybala@pointblue.org}
}
\references{
Brand LA, Takekawa JY, Shinn J, Graham T, Buffington K, Spring S, Miles K. 2013. Effects of wetland management on carrying capacity of duck and shorebird benthivores in a coastal estuary. Vallejo, CA: U.S. Geological Survey, Western Ecological Research Center.

Castro G, Stoyan N, Myers JP. 1989. Assimilation efficiency in birds: A function of taxon or food type? Comp. Biochem. Physiol. 92A:271-278.

Kersten M, Piersma T. 1987. High levels of energy expenditure in shorebirds; Metabolic adaptations to an energetically expensive way of life. Ardea 75:175-187.
}

